add location tracker

+++ app/src/main/java/com/example/senla_tz/entify/Track.kt
@@ -11,5 +11,5 @@ data class Track(
     var idServer: Long,
     var beginsAt: Long,
     var time: Long,
-    var distance: Long
+    var distance: Int
 )
